Restructure and expand test coverage
