+++ app/test_objects/test_controller.rb
@@ -1,0 +1,4 @@
+class TestController < UIViewController
+  include DelegateTestAttributes
+end
+
