+++ app/test_objects/test_screen.rb
@@ -1,0 +1,4 @@
+class TestScreen < PM::Screen
+  include DelegateTestAttributes
+  stylesheet TestScreenStylesheet
+end
