+++ app/test_objects/test_screen_stylesheet.rb
@@ -1,0 +1,9 @@
+class TestScreenStylesheet < ApplicationStylesheet
+  def root_view(st)
+    st.background_color = color.white
+  end
+  def test_label(st)
+    st.text = 'style from sheet'
+  end
+end
+
