+++ spec/helpers/cdq.rb
@@ -1,4 +1,3 @@
-
 module Bacon
   class Context
     include CDQ
