--- spec/screens/metal_table_screen_spec.rb
@@ -1,8 +1,0 @@
-describe 'MetalTableScreen' do
-
-  before do
-  end
-
-  after do
-  end
-end
